sstatesig: Ensure spdx dependencies update when abisafe recipes change
If you build openssh, bump the version of base-passwd, then rerun openssh:do_create_spdx,
it will fail due to references to the old version of base-passwd. This is due to
base-passwd being listed in SIGGEN_EXCLUDERECIPES_ABISAFE.

Ignore these recipes for the purposes of create_spdx and force rebuilds when things change.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>