qemu: add PACKAGECONFIG for sndio
In qemu's configuration this is an 'auto' configuration option so if the host has
the sndio headers on then it gets enabled, causing sstate reuse problems. This is
particularly problematic as some qemu pieces can 'see' into the host pkg-config
due to openGL linkage.

Add config and disable by default to avoid host contamination.

[YOCTO #15284]

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>